+++ libs/fn/src/lib/cdk/interfaces/constructor.type.ts
