+++ libs/fn/src/lib/cdk/mixins/can-assign-additional-classes.ts
@@ -1,4 +1,4 @@
-import { Constructor } from '../interfaces/Constructor';
+import { Constructor } from '../interfaces/constructor.type';
 import { Directive, HostBinding, Input } from '@angular/core';
 
 export interface HasAdditionalClasses {
