+++ libs/fn/src/lib/cdk/mixins/has-tab-index.ts
@@ -1,4 +1,4 @@
-import { Constructor } from '../interfaces/Constructor';
+import { Constructor } from '../interfaces/constructor.type';
 import { Directive, HostBinding, Input } from '@angular/core';
 import { CanBeDisabledConstructor } from './can-be-disabled';
